pyproject.toml
[project]
name = "sentiflow"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.6.2",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi==0.136.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "markitdown>=0.1.3",
    "pydantic-settings>=2.14.0",
    "python-multipart>=0.0.20",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "uvicorn==0.46.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
